Reorganized components into pages and uicomponents folders. This should make it more clear which components are core functionality and pages to the app, and which are things like jumbotron and other bootstrap UI
