+++ client/src/Components/Pages/ProfileNav/index.js
@@ -1,6 +1,6 @@
 import React from "react";
 import "./style.css";
-import { Input, TextArea, FormBtn,FormBtnlink } from "../Form";
+import { Input, TextArea, FormBtn,FormBtnlink } from "../../UiComponents/Form";
 // import categories from "./categories.json";
 import {Link} from "react-router-dom";
