--- client/src/Components/main/index.js
@@ -1,10 +1,0 @@
-import React from "react";
-import "./style.css";
-
-function main() {
-  return (
-    <div className="container"></div>
-  );
-}
-
-export default main;
