--- client/src/Components/main/style.css
@@ -1,4 +1,0 @@
-.container{
-    background-color: #e74b3c;
-    height:100%;
-}
